MDL-20636 Another little bug found by Jamie.

+++ mod/quiz/locallib.php
@@ -748,8 +748,8 @@ function quiz_update_all_final_grades($quiz) {
 
     if (!empty($todelete)) {
         list($test, $params) = $DB->get_in_or_equal($todelete);
-        $DB->delete_records_select('quiz_grades',
-                'quiz = ? AND userid ', array($quiz->id) + $params);
+        $DB->delete_records_select('quiz_grades', 'quiz = ? AND userid ' . $test,
+                array_merge(array($quiz->id), $params));
     }
 }
